4.15.-4.23.

+++ part4/bloglist/controllers/blogs.js
@@ -1,19 +1,43 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user',
+      {
+        username: 1,
+        name: 1
+      }
+    )
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const body = request.body
+  const user = request.user
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user.id
+  })
 
   const addedBlog = await blog.save()
+  user.blogs = user.blogs.concat(addedBlog._id)
+  await user.save()
   response.status(201).json(addedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== request.user.id.toString()) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
